convert to C++ replace lame config parser with two-pass parser allowing rehash
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@17089 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ willow/src/include/wcache.h
@@ -36,7 +36,7 @@ struct cache_object {
 	time_t		  co_time;	/* Last-Modified / retrieval time	*/
 	time_t		  co_lru;	/* Last used timestamp, for eviction	*/
 	int		  co_id;	/* Object id				*/
-	int		  co_plen;	/* Size of cache object path		*/
+	size_t		  co_plen;	/* Size of cache object path		*/
 	size_t		  co_size;	/* Object size				*/
 	char		 *co_path;	/* Object data location			*/
 	int		  co_complete;	/* Finished being cached		*/
